Fixed a number of issues regarding the naming of constant parameters in the timed semantics.
Changed various semantic functions to carry the set of defined
operations. This is required to calculate the semantics of Call
statements of defined operations, which need to pass constants which are
parameters of the state machine.

Fixed naming conventions for global elements (functions and types) to
use the package name. In this way, it is possible to define, for
instance, multiple types with the same in different packages and still
generate the semantics.

Changed calculate of assertions file to include all _defs if the
instantiations.csp file is used. This is needed because the
instantiations.csp file is unique to a project and could use data
defined in a package not directly required by the assertions file.
If the assertions file defines its own instantiations block, the
definitions are not included as they are not necessary and would
otherwise force the definition of all constants, functions etc.

Added AutonomousChemicalDetector to the tests projects. It now passes
all tests.