Added project that exercises multiple parallel calls of same operation.
This is done for both RoboChart and RoboSim. Interestingly, in RoboSim
because the choice of outputs is made internally by the machine in a
controller, then the controller is non-deteterministic. Equally, the
RoboChart model when constrained by the assumptions for simulation also
makes it non-deterministic.

We also check for conformance between models and that specific traces
with duplicated outputs are absent in the models.